+++ src/main/java/frc/robot/subsystems/Lift.java
@@ -10,33 +10,13 @@ import frc.robot.commands.*;
 /**
  * 
  */
-public class Lift extends Subsystem {
-
-    // Put methods for controlling this subsystem
-    // here. Call these from Commands.
-    
-    //Lift for Monolith. Should work.
-
-    Talon motor = new Talon(RobotMap.liftPortMonolith);
-    
-    DigitalInput topLimit = new DigitalInput(1);
-    DigitalInput bottomLimit = new DigitalInput(2);
-
+public abstract class Lift extends Subsystem {
     public void initDefaultCommand() {
         setDefaultCommand(new LiftControl());
     }
     
-    public void moveLift(double v)
-    {
-        if ((v < 0 & bottomLimit.get()) || (v > 0 & topLimit.get()))
-        {
-            motor.set(v);
-        }
-    }
-    
-    public void stop()
-    {
-    	motor.stopMotor();
-    	motor.set(0.0);
-    }
+    public abstract void moveLift(double v);
+
+    public abstract void stop();
+
 }
